Fixed bug where app wont archive if BitCode is enabled
Discovered that the plugin caused an issue where the application it was added to would fail to archive because the libraries were made without BitCode being enabled. This prevented the app from being used with TestFlight or Unity Cloud Build. Rebuilt libraries with BitCode enabled. Will need to update build scripts in library projects to copy the library to the correct unity folders, they do not appear to work when the library is build via the "archive" command.